Complete Shipment Updater Class
Read in from S3::Objects and update relevant information for shipments
that Fosdick has shipped. Overall it may make more sense to dump this
into a queue using Active Job but I fear many of the Spree stores that
use Fosdick might not be that up to date with their Rails ecosystem.