Add Dotenv Rails For Environment Vars
Environment vars will be required when testing that items are properly
uploaded to the S3 Buckets. When writing the e2e style test we will need
to assert that a file is properly uploaded and downloaded/read by the
application. Dotenv will be used to expose these vars to the test
runner/environment.