Pending Upload Test.
Documentation for really what the test needs to assert. Moving to a tag
system as to be able to filter out the slow tests. This is over wiring
up VCR for now. The uploader and downloader might have 4 to 5 total
tests and it may not be the best to have them wired up to cassettes.