Factory Pattern For Uploading A Document
Due to the use of this gem being predominantly through rake tasks, move
most logic into a single entry point for logic. This is basically the
factory pattern so incorporate this in the uploading of documents. Old
shipment upload spec moved to factory spec. Green tests! Document is
uploaded successfully to s3