Clean Up Uploader And Spec
The uploader and spec will change a little bit due to the factory that
uses this class to upload items to S3. This class is a thin wrapper
around the S3 gem that allows initiation with a bucket and a single
public upload method that takes some content and a file name (both
strings). This then uploads the content to the S3 bucket with the name
that is passed to the upload method.