Scaffolds Extraction of IndividualShipmentUpdater
It would be nice to extract the behaviour in ShipmentsUpdater that updates each
individual shipment, so that specific integrations could augment its behaviour
with less impact. This commit adds a set of pending specs that describe the
behaviour of ShipmentsUpdater#update_shipment such that it can be refactored
into IndividualShipmentUpdater, and then ultimately refactored further.